+++ src/container/OtherContainer/Queue.ts
@@ -1,5 +1,5 @@
-import { Base, initContainer } from '@/container/ContainerBase/index';
 import Deque from '../SequentialContainer/Deque';
+import { Base, initContainer } from '@/container/ContainerBase/index';
 
 class Queue<T> extends Base {
   private queue: Deque<T>;