+++ src/container/SequentialContainer/Vector.ts
@@ -1,70 +1,9 @@
-import { ContainerIterator, initContainer } from '@/container/ContainerBase/index';
-import { checkWithinAccessParams } from '@/utils/checkParams';
 import SequentialContainer from './Base/index';
+import { checkWithinAccessParams } from '@/utils/checkParams';
+import { ContainerIterator, initContainer } from '@/container/ContainerBase/index';
+import { RandomIterator } from '@/container/SequentialContainer/Base/RandomIterator';
 
-export class VectorIterator<T> extends ContainerIterator<T> {
-  private node;
-  private readonly size: () => number;
-  private readonly getElementByPos: (pos: number) => T;
-  private readonly setElementByPos: (pos: number, element: T) => void;
-  pre: () => this;
-  next: () => this;
-  constructor(
-    index: number,
-    size: () => number,
-    getElementByPos: (pos: number) => T,
-    setElementByPos: (pos: number, element: T) => void,
-    iteratorType?: boolean
-  ) {
-    super(iteratorType);
-    this.node = index;
-    this.size = size;
-    this.getElementByPos = getElementByPos;
-    this.setElementByPos = setElementByPos;
-
-    if (this.iteratorType === ContainerIterator.NORMAL) {
-      this.pre = function () {
-        if (this.node === 0) {
-          throw new RangeError('Deque iterator access denied!');
-        }
-        this.node -= 1;
-        return this;
-      };
-      this.next = function () {
-        if (this.node === this.size()) {
-          throw new RangeError('Deque Iterator access denied!');
-        }
-        this.node += 1;
-        return this;
-      };
-    } else {
-      this.pre = function () {
-        if (this.node === this.size() - 1) {
-          throw new RangeError('Deque iterator access denied!');
-        }
-        this.node += 1;
-        return this;
-      };
-      this.next = function () {
-        if (this.node === -1) {
-          throw new RangeError('Deque iterator access denied!');
-        }
-        this.node -= 1;
-        return this;
-      };
-    }
-  }
-  get pointer() {
-    checkWithinAccessParams(this.node, 0, this.size() - 1);
-    return this.getElementByPos(this.node);
-  }
-  set pointer(newValue: T) {
-    checkWithinAccessParams(this.node, 0, this.size() - 1);
-    this.setElementByPos(this.node, newValue);
-  }
-  equals(obj: VectorIterator<T>) {
-    return this.node === obj.node;
-  }
+export class VectorIterator<T> extends RandomIterator<T> {
   copy() {
     return new VectorIterator(
       this.node,