+++ test/HashContainerTest/HashSet.test.ts
@@ -1,5 +1,4 @@
-import HashContainerBase from '@/container/HashContainer/Base/index';
-import { Vector, HashSet } from '@/index';
+import { Vector, HashSet, HashContainer } from '@/index';
 
 function generateRandom(low = 0, high = 1e6, fix = 6) {
   return (low + Math.random() * (high - low)).toFixed(fix);
@@ -20,9 +19,9 @@ function judgeHashSet(myHashSet: HashSet<string>, stdSet: Set<string>) {
 
 describe('HashSet test', () => {
   // @ts-ignore
-  HashContainerBase.treeifyThreshold = 1;
+  HashContainer.treeifyThreshold = 1;
   // @ts-ignore
-  HashContainerBase.untreeifyThreshold = 1;
+  HashContainer.untreeifyThreshold = 1;
   test('constructor test', () => {
     // eslint-disable-next-line no-new
     expect(() => new HashSet([], 28)).toThrow(RangeError);
@@ -88,7 +87,7 @@ describe('HashSet test', () => {
     // @ts-ignore
     const bucketNum = myHashSet.bucketNum;
     // @ts-ignore
-    myHashSet.bucketNum = HashContainerBase.maxBucketNum;
+    myHashSet.bucketNum = HashContainer.maxBucketNum;
     // @ts-ignore
     myHashSet.reAllocate();
     // @ts-ignore
@@ -101,5 +100,64 @@ describe('HashSet test', () => {
     myHashSet.bucketNum = bucketNum;
     // eslint-disable-next-line no-unused-vars, @typescript-eslint/no-unused-vars, no-empty
     for (const _ of myHashSet) {}
+  });
+
+  test('HashSet normal test', () => {
+    // @ts-ignore
+    HashContainer.treeifyThreshold = 6;
+    // @ts-ignore
+    HashContainer.untreeifyThreshold = 8;
+    const st = new HashSet<string>();
+    const stdSet = new Set<string>();
+    for (let i = 0; i < testNum; ++i) {
+      st.insert(i.toString());
+      stdSet.add(i.toString());
+    }
+    judgeHashSet(st, stdSet);
+    let size = testNum;
+    for (let i = 0; i < testNum; ++i) {
+      st.eraseElementByKey(i.toString());
+      expect(st.size()).toEqual(--size);
+    }
+    expect(st.size()).toEqual(0);
+  });
+
+  test('HashSet insert and erase', () => {
+    // @ts-ignore
+    HashContainer.treeifyThreshold = 6;
+    // @ts-ignore
+    HashContainer.untreeifyThreshold = 8;
+    const st = new HashSet<string>();
+    const stdSet = new Set<string>();
+    const arr: string[] = [];
+    for (let i = 0; i < testNum; ++i) {
+      const random = Math.random().toFixed(6);
+      st.insert(random);
+      stdSet.add(random);
+      arr.push(random);
+    }
+    judgeHashSet(st, stdSet);
+    for (let i = 0; i < testNum; ++i) {
+      if (Math.random() > 0.5) {
+        st.eraseElementByKey(arr[i]);
+        stdSet.delete(arr[i]);
+      }
+    }
+    judgeHashSet(st, stdSet);
+    arr.length = 0;
+    for (let i = 0; i < testNum; ++i) {
+      const random = Math.random().toFixed(6);
+      st.insert(random);
+      stdSet.add(random);
+      arr.push(random);
+    }
+    judgeHashSet(st, stdSet);
+    for (let i = 0; i < testNum; ++i) {
+      if (Math.random() > 0.5) {
+        st.eraseElementByKey(arr[i]);
+        stdSet.delete(arr[i]);
+      }
+    }
+    judgeHashSet(st, stdSet);
   });
 });