+++ test/OtherTest/iterator.test.ts
@@ -4,13 +4,16 @@ import {
   Deque,
   OrderedSet,
   OrderedMap,
+  Container,
+  ContainerIterator,
   VectorIterator,
   LinkListIterator,
   DequeIterator,
   OrderedSetIterator,
-  OrderedMapIterator
+  OrderedMapIterator,
+  SequentialContainer,
+  TreeContainer
 } from '@/index';
-import { ContainerIterator } from '@/container/ContainerBase';
 
 let arr: number[] = [];
 const testNum = 10000;
@@ -20,7 +23,7 @@ for (let i = 0; i < testNum; ++i) {
 arr = Array.from(new Set(arr));
 arr.sort((x, y) => x - y);
 
-const containerArr = [
+const containerArr: Container<unknown>[] = [
   new Vector(arr),
   new LinkList(arr),
   new Deque(arr),
@@ -157,6 +160,18 @@ describe('iterator test', () => {
         }, function () {
           return 0;
         });
+      // @ts-ignore
+      // eslint-disable-next-line no-new
+      new Container();
+      // @ts-ignore
+      // eslint-disable-next-line no-new
+      new ContainerIterator();
+      // @ts-ignore
+      // eslint-disable-next-line no-new
+      new SequentialContainer();
+      // @ts-ignore
+      // eslint-disable-next-line no-new
+      new TreeContainer();
       // @ts-ignore
       // eslint-disable-next-line no-new
       new LinkListIterator<number>(containerArr[0].begin().node);