+++ CHANGELOG.md
@@ -4,11 +4,28 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
-## [UnReleased] - 2022.09.04
+## [UnReleased] - 2022.09.08
 
 ### Added
 
 - Add `index` property to tree iterator which represents the sequential index of the iterator in the tree.
+
+## [4.1.4] - 2022.09.07
+
+### Added
+
+- Add some notes.
+
+### Changed
+
+- Optimize hash container.
+- Abstracting out the hash container.
+
+### Fixed
+
+- Fixed tree get height function return one larger than the real height.
+- Tree-shaking not work in ES module.
+- `Queue` and `Deque` should return `undefined` when container is empty.
 
 ## [4.1.4-beta.0] - 2022.08.31